new stats design and fixes

+++ application/src/main/java/nl/jixxed/eliteodysseymaterials/domain/ShipConfigurationModification.java
@@ -5,11 +5,14 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import nl.jixxed.eliteodysseymaterials.enums.HorizonsBlueprintGrade;
 import nl.jixxed.eliteodysseymaterials.enums.HorizonsBlueprintType;
+
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 public class ShipConfigurationModification {
     private HorizonsBlueprintType type;
     private HorizonsBlueprintGrade grade;
-    private Double percentComplete;
+    private BigDecimal percentComplete;
 }
